Created a new SubArticle file to remove the blank article when we selected the submit button to add an article
