+++ src/scripts/articles.js
@@ -12,8 +12,7 @@ const article = Object.create({}, {
             userHome.buildUserHome();
             const secRef = document.querySelector("#userContentSection");
             const addArticleButton = $("<button>").attr("id", "addArticle").appendTo(secRef).text("Add Article")
-                //needs to add article by article id
-            const $article = $("<article>").attr("id", "idValue").appendTo(secRef)
+
 
         }
     }
